requirements.txt
streamlit==1.35.0
pandas==2.2.3
tensorflow-cpu==2.17.1
pillow==11.1.0
matplotlib==3.9.2
transformers==4.38.2
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
numpy==1.26.4
scikit-learn==1.4.2
plotly==5.22.0
